refactor: rebuild folder tructure of event detail
