+++ frontend/app/components/event-detail/sections/EventAds.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+defineProps<{
+  sideBarImage: string
+}>();
+</script>
+
+<template>
+    <div class="col-lg-3 mt-3 d-none d-lg-block px-2 ">
+        <img :src="sideBarImage" alt="" class="rounded-5 img-fluid object-fit-cover"/>
+    </div>
+</template>
