+++ frontend/i18n/locales/en.json
@@ -60,7 +60,7 @@
     "shipping_terms": "Shipping Terms",
     "payment_methods": "Payment Methods"
   },
-  "even-detail": {
+  "event-detail": {
     "price_from": "From",
     "buy_tickets": "Buy Tickets Now",
     "schedule_event": "Event Schedule",
